Move ephemeris routines to a new file.
