+++ src/Makefile
@@ -127,6 +127,7 @@ CSRC := $(PORTSRC) \
         $(SWIFTNAV_ROOT)/src/nmea.o \
         $(SWIFTNAV_ROOT)/src/system_monitor.o \
         $(SWIFTNAV_ROOT)/src/flash_callbacks.o \
+        $(SWIFTNAV_ROOT)/src/ephemeris.o \
         main.c
 
 # C++ sources that can be compiled in ARM or THUMB mode depending on the global
