+++ src/ephemeris.h
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2015 Swift Navigation Inc.
+ * Contact: Fergus Noble <fergus@swift-nav.com>
+ *          Gareth McMullin <gareth@swiftnav.com>
+ *
+ * This source is subject to the license found in the file 'LICENSE' which must
+ * be be distributed together with this source. All other rights reserved.
+ *
+ * THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,
+ * EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
+ */
+#ifndef SWIFTNAV_EPHEMERIS_H
+#define SWIFTNAV_EPHEMERIS_H
+
+#include <libswiftnav/constants.h>
+#include <libswiftnav/ephemeris.h>
+
+extern ephemeris_t es[MAX_SATS];
+
+void ephemeris_setup(void);
+
+#endif
+
