+++ src/main.c
@@ -35,6 +35,7 @@
 #include "simulator.h"
 #include "settings.h"
 #include "sbp_fileio.h"
+#include "ephemeris.h"
 
 extern void ext_setup(void);
 
@@ -42,74 +43,10 @@ extern void ext_setup(void);
 #define SYSTEM_CLOCK 130944000
 #endif
 
-/* TODO: Think about thread safety when updating ephemerides. */
-ephemeris_t es[32] _CCM;
-ephemeris_t es_old[32] _CCM;
-
 /* Required by exit() which is called from BLAS/LAPACK. */
 void _fini(void)
 {
   return;
-}
-
-static WORKING_AREA_CCM(wa_nav_msg_thread, 3000);
-static msg_t nav_msg_thread(void *arg)
-{
-  (void)arg;
-  chRegSetThreadName("nav msg");
-
-  memset(es, 0, sizeof(es));
-  for (u8 i=0; i<32; i++) {
-    es[i].prn = i;
-  }
-
-  while (TRUE) {
-
-    /* TODO: This should be trigged by a semaphore from the tracking loop, not
-     * just ran periodically. */
-
-
-    for (u8 i=0; i<nap_track_n_channels; i++) {
-      chThdSleepMilliseconds(100);
-      /* Check if there is a new nav msg subframe to process.
-       * TODO: move this into a function */
-      if (tracking_channel[i].state == TRACKING_RUNNING &&
-          tracking_channel[i].nav_msg.subframe_start_index) {
-
-        /* Save old ephemeris before potentially updating. */
-        memcpy(&es_old[tracking_channel[i].prn],
-               &es[tracking_channel[i].prn],
-               sizeof(ephemeris_t));
-
-        __asm__("CPSID i;");
-        s8 ret = process_subframe(&tracking_channel[i].nav_msg,
-                                  &es[tracking_channel[i].prn]);
-        __asm__("CPSIE i;");
-
-        if (ret < 0) {
-          log_info("PRN %02d ret %d\n", tracking_channel[i].prn+1, ret);
-        } else if (ret == 1) {
-          /* Decoded a new ephemeris. */
-
-          if (memcmp(&es[tracking_channel[i].prn],
-                     &es_old[tracking_channel[i].prn],
-                     sizeof(ephemeris_t))) {
-            log_info("New ephemeris for PRN %02d\n", tracking_channel[i].prn+1);
-          }
-
-          if (!es[tracking_channel[i].prn].healthy) {
-            log_info("PRN %02d unhealthy\n", tracking_channel[i].prn+1);
-          } else {
-            sbp_send_msg(SBP_MSG_EPHEMERIS,
-                         sizeof(ephemeris_t),
-                         (u8 *)&es[tracking_channel[i].prn]);
-          }
-        }
-      }
-    }
-  }
-
-  return 0;
 }
 
 /** Compare version strings.
@@ -279,8 +216,7 @@ int main(void)
                       TYPE_INT);
   READ_ONLY_PARAMETER("system_info", "nap_fft_index_bits", nap_acq_fft_index_bits, TYPE_INT);
 
-  chThdCreateStatic(wa_nav_msg_thread, sizeof(wa_nav_msg_thread),
-                    NORMALPRIO-1, nav_msg_thread, NULL);
+  ephemeris_setup();
 
   /* Send message to inform host we are up and running. */
   u32 startup_flags = 0;
